refactor(ucmc-web): regroup /my/account sections by sensitivity
Move email management onto the Sign-in tab (renamed from Security)
next to passkeys, leaving Details focused on private PII (name, phone,
emergency contacts). Relocate the data-export and account-deletion
controls to Preferences so destructive account-wide actions sit apart
from credential management. Update the Sign-in loader to prefetch
both passkeys and emails, and add a cross-link from Details so users
following the old mental model can still find the email list.

Switch the email rows to shadcn Item/ItemGroup primitives for visual
consistency with the rest of the account surface, and fix the mobile
tab strip to scroll horizontally instead of wrapping onto two lines.
