+++ apps/web/src/routes/my.account.security.tsx
@@ -1,31 +1,24 @@
 import { useQuery } from "@tanstack/react-query";
-import { createFileRoute, useNavigate } from "@tanstack/react-router";
-import { useState } from "react";
-import { toast } from "sonner";
+import { createFileRoute, Link } from "@tanstack/react-router";
 
-import { passkeyListQueryOptions } from "#/features/auth/api/queries";
-import { useDeleteMyAccount } from "#/features/auth/api/use-delete-my-account";
+import {
+  myEmailsQueryOptions,
+  passkeyListQueryOptions,
+} from "#/features/auth/api/queries";
 import { useRemovePasskey } from "#/features/auth/api/use-remove-passkey";
 import { useAuth } from "#/features/auth/api/use-auth";
+import { EmailAddressesSection } from "#/features/auth/components/email-addresses-section";
 import { AddPasskeyButton } from "#/features/auth/components/passkey-button";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "#/components/ui/alert-dialog";
 import { Button } from "#/components/ui/button";
-import { Input } from "#/components/ui/input";
-import { Label } from "#/components/ui/label";
 
 /**
- * Account → Security. Lists the user's registered passkeys with an
- * "Add this device" affordance and a per-row "Remove" button.
+ * Account → Sign-in. Surfaces the two ways a member authenticates:
+ * verified email addresses (magic-link targets) and registered
+ * passkeys. The tab URL is still `/my/account/security` — only the
+ * label changed — so existing bookmarks keep working.
+ *
+ * Data export and account deletion moved to the Preferences tab; those
+ * are privacy/account-management concerns rather than sign-in surface.
  *
  * Auth-gating is inherited from the parent `/my` layout (`my.tsx`),
  * which runs `requireApproved` for the entire `/my/*` namespace —
@@ -33,29 +26,42 @@ import { Label } from "#/components/ui/label";
  */
 export const Route = createFileRoute("/my/account/security")({
   loader: async ({ context }) => {
-    await context.queryClient.ensureQueryData(passkeyListQueryOptions());
+    await Promise.all([
+      context.queryClient.ensureQueryData(passkeyListQueryOptions()),
+      context.queryClient.ensureQueryData(myEmailsQueryOptions()),
+    ]);
   },
-  component: SecurityPage,
+  component: SignInPage,
 });
 
-function SecurityPage() {
+function SignInPage() {
+  const { principal } = useAuth();
   const query = useQuery(passkeyListQueryOptions());
   const removal = useRemovePasskey();
 
   const passkeys = query.data ?? [];
 
+  if (!principal) {
+    return null;
+  }
+
   return (
     <div className="flex flex-col gap-8">
-      <header className="space-y-2">
-        <p className="text-sm text-muted-foreground">
-          Passkeys let you sign in with Face ID, Touch ID, a Windows Hello PIN,
-          or a hardware security key. Register one on every device you use and
-          you&rsquo;ll never need the emailed sign-in link again.
-        </p>
-      </header>
+      <EmailAddressesSection
+        approved={principal.status === "approved"}
+        primaryEmail={principal.primaryEmail}
+      />
 
       <section className="space-y-3">
-        <h2 className="text-lg font-semibold">Your passkeys</h2>
+        <header>
+          <h3 className="text-base font-medium">Passkeys</h3>
+          <p className="text-sm text-muted-foreground">
+            Passkeys let you sign in with Face ID, Touch ID, a Windows Hello
+            PIN, or a hardware security key. Register one on every device you
+            use and you&rsquo;ll never need the emailed sign-in link again.
+          </p>
+        </header>
+
         {query.isLoading ? (
           <p className="text-sm text-muted-foreground">Loading…</p>
         ) : passkeys.length === 0 ? (
@@ -104,126 +110,17 @@ function SecurityPage() {
             ))}
           </ul>
         )}
-      </section>
 
-      <section className="space-y-3">
-        <h2 className="text-lg font-semibold">Add a passkey</h2>
         <AddPasskeyButton />
       </section>
 
-      <DataAndDeletionSection />
+      <p className="text-xs text-muted-foreground">
+        Looking to download your data or delete your account? Those live on the{" "}
+        <Link to="/my/account/preferences" className="underline">
+          Preferences
+        </Link>{" "}
+        tab.
+      </p>
     </div>
-  );
-}
-
-/**
- * Data export + account deletion. Both honor the promises on /privacy:
- * download a JSON copy of everything we have on you, and hard-delete
- * the account immediately. Deletion is gated by typing the email on
- * file — defends against accidental clicks more than it does against
- * deliberate misuse (a hostile actor with the session can already type
- * the email).
- */
-function DataAndDeletionSection() {
-  const { principal } = useAuth();
-  const navigate = useNavigate();
-  const deletion = useDeleteMyAccount();
-  const [open, setOpen] = useState(false);
-  const [confirmEmail, setConfirmEmail] = useState("");
-
-  const expectedEmail = principal?.primaryEmail ?? "";
-  const matches = confirmEmail.trim().toLowerCase() === expectedEmail;
-
-  const onConfirm = () => {
-    deletion.mutate(undefined, {
-      onSuccess: async () => {
-        toast.success("Account deleted");
-        setOpen(false);
-        await navigate({ to: "/" });
-      },
-      onError: (err) => {
-        toast.error(
-          err instanceof Error ? err.message : "Couldn't delete the account",
-        );
-      },
-    });
-  };
-
-  return (
-    <>
-      <section className="space-y-3">
-        <h2 className="text-lg font-semibold">Your data</h2>
-        <p className="text-sm text-muted-foreground">
-          Download a JSON copy of everything this site stores about you — your
-          profile, emergency contacts, role memberships, and waiver attestation
-          history. The export does not include passkey credentials or magic-link
-          tokens.
-        </p>
-        <Button asChild variant="outline">
-          <a href="/api/account/export" download>
-            Download my data
-          </a>
-        </Button>
-      </section>
-
-      <section className="space-y-3">
-        <h2 className="text-lg font-semibold text-destructive">
-          Delete account
-        </h2>
-        <p className="text-sm text-muted-foreground">
-          Permanently delete your account, profile, emergency contacts,
-          passkeys, sessions, and waiver attestation history. This is immediate
-          and irreversible — there is no recovery on the other side. Authored
-          announcements stay (anonymized) so the timeline for other members
-          remains intact.
-        </p>
-        <AlertDialog open={open} onOpenChange={setOpen}>
-          <AlertDialogTrigger asChild>
-            <Button variant="destructive">Delete my account</Button>
-          </AlertDialogTrigger>
-          <AlertDialogContent>
-            <AlertDialogHeader>
-              <AlertDialogTitle>Delete your account?</AlertDialogTitle>
-              <AlertDialogDescription>
-                This is permanent. Type{" "}
-                <strong className="font-mono">{expectedEmail}</strong> to
-                confirm.
-              </AlertDialogDescription>
-            </AlertDialogHeader>
-            <div className="space-y-2 py-2">
-              <Label htmlFor="confirm-delete-email" className="text-sm">
-                Email
-              </Label>
-              <Input
-                id="confirm-delete-email"
-                type="email"
-                autoComplete="off"
-                value={confirmEmail}
-                onChange={(e) => setConfirmEmail(e.target.value)}
-                placeholder={expectedEmail}
-              />
-            </div>
-            <AlertDialogFooter>
-              <AlertDialogCancel
-                onClick={() => setConfirmEmail("")}
-                disabled={deletion.isPending}
-              >
-                Cancel
-              </AlertDialogCancel>
-              <AlertDialogAction
-                onClick={(e) => {
-                  e.preventDefault();
-                  onConfirm();
-                }}
-                disabled={!matches || deletion.isPending}
-                className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
-              >
-                {deletion.isPending ? "Deleting…" : "Delete forever"}
-              </AlertDialogAction>
-            </AlertDialogFooter>
-          </AlertDialogContent>
-        </AlertDialog>
-      </section>
-    </>
   );
 }
